Fix update of external line number when cleared
This fix corrects a minor cosmetic issue that occurs when a player attempts to clear an existing line number value (1-9) of a wagon.  Specifically, clearing the line number field via the onboard computer in the driver's stand does not immediately cause the external display of the line number to be cleared.  It will be cleared eventually if the player triggers a reload of the wagon such as when leaving and then returning to the area of the wagon or by exiting and restarting the game.  This fix causes the external line number to be cleared immediately without needing to trigger a reload of the wagon.